requirements.txt
pandas
numpy
nltk
textblob
gcloud
google-cloud-storage